Refactor charmlite to avoid cyclic dependencies between headers
This commit provides the following changes:
1. Decouple common.hpp such that types instantiation of incomplete
types now occur in their respective headers.
2. Split message.hpp to message and impl. The implementation
relies on destination and the respective free functions are
defined there as well.
3. Split proxy.hpp to proxy and impl. The implementation of some
functions introduced cyclic dependency with completion and
collection.
4. Core related functions have been transferred to charmlite.hpp
and so has the functions from charmlite.impl.hpp.

Random Flyby:
1. Change enum to enum class
2. Implement meta-template solution to is_message
