+++ examples/cbench/Makefile
@@ -5,7 +5,7 @@ OPTS:=$(OPTS) -Ofast
 all: pgm
 
 pgm: pgm.o ../../libs/core.o
-	$(CXX) $(OPTS) ../../libs/core.o pgm.o -o pgm
+	$(CXX) $(OPTS) pgm.o ../../libs/core.o -o pgm
 
 pgm.o: pgm.cpp
 	$(CXX) $(OPTS) -c -o pgm.o pgm.cpp
