+++ include/charmlite/core/options.hpp
@@ -1,5 +1,5 @@
-#ifndef __CMK_OPTIONS_HH__
-#define __CMK_OPTIONS_HH__
+#ifndef CHARMLITE_CORE_OPTIONS_HPP
+#define CHARMLITE_CORE_OPTIONS_HPP
 
 #include <charmlite/core/common.hpp>
 
@@ -111,6 +111,7 @@ namespace cmk {
             return converter::decode(this->step_);
         }
     };
+
 }    // namespace cmk
 
 #endif
