+++ include/charmlite/utilities/traits.hpp
@@ -1,7 +1,7 @@
-#ifndef __CMK_TRAITS_HH__
-#define __CMK_TRAITS_HH__
+#ifndef CHARMLITE_UTILITIES_TRAITS_HPP
+#define CHARMLITE_UTILITIES_TRAITS_HPP
 
-#include <charmlite/core/message.hpp>
+#include <charmlite/utilities/traits/message.hpp>
 
 namespace cmk {
 
@@ -20,20 +20,6 @@ namespace cmk {
     template <typename T>
     using viable_argument_t = typename viable_argument<T>::type;
 
-    template <typename T, typename Message>
-    using member_fn_t = void (T::*)(cmk::message_ptr<Message>&&);
-
-    template <typename T>
-    constexpr bool is_message_(void)
-    {
-        return std::is_base_of<message, T>::value;
-    }
-
-    template <>
-    constexpr bool is_message_<void>(void)
-    {
-        return true;
-    }
 }    // namespace cmk
 
 #endif
